requirements.txt
flask-restplus==0.13.0
flask-migrate==2.7.0
flask-script==2.0.6
flask-testing==0.8.1
sqlalchemy==1.4.5
flask-sqlalchemy==2.5.1
psycopg2==2.8.6
Werkzeug==0.16.1
# Development
autopep8==1.5.6
ipdb==0.13.7